Framebuffer work:
- can select book from list on framebuffer (sdl; kobo still needs events wired)
- improved font sizing on framebuffer
- early work on windowing and event systems
- unit testing
- thinking on FontEngine abstraction

+++ ocher/ux/fb/RenderFb.h
@@ -25,7 +25,7 @@ protected:
     int m_col;
     int m_penX;
     int m_penY;
-    int m_lineHeight;
+    uint8_t m_lineHeight;
 
     void pushAttrs();
     void applyAttrs(int i);
